Breakout the generator methods into their own package.
